feat: use primordial git HEAD
`usize::from(a) as _` can be transformed to
`a.try_into().or(Err(libc::EINVAL))?`

with the not yet released primordial 0.2

Because the `sgx` crate still uses primordial 0.1,
`internal/shim-sgx/src/event.rs` and `internal/shim-sgx/src/handler.rs`
are force transmuted as a workaround for now.

+++ internal/shim-sgx/src/handler.rs
@@ -80,7 +80,7 @@ impl<'a> Handler<'a> {
             );
         }
 
-        self.block.msg.req = request!(SYS_ENARX_CPUID => self.aex.gpr.rax, self.aex.gpr.rcx);
+        self.block.msg.req = request!(SYS_ENARX_CPUID => usize::from(self.aex.gpr.rax), usize::from(self.aex.gpr.rcx));
 
         unsafe {
             // prevent earlier writes from being moved beyond this point
@@ -91,10 +91,10 @@ impl<'a> Handler<'a> {
             // prevent later reads from being moved before this point
             core::sync::atomic::compiler_fence(core::sync::atomic::Ordering::Acquire);
 
-            self.aex.gpr.rax = self.block.msg.req.arg[0].into();
-            self.aex.gpr.rbx = self.block.msg.req.arg[1].into();
-            self.aex.gpr.rcx = self.block.msg.req.arg[2].into();
-            self.aex.gpr.rdx = self.block.msg.req.arg[3].into();
+            self.aex.gpr.rax = usize::from(self.block.msg.req.arg[0]).into();
+            self.aex.gpr.rbx = usize::from(self.block.msg.req.arg[1]).into();
+            self.aex.gpr.rcx = usize::from(self.block.msg.req.arg[2]).into();
+            self.aex.gpr.rdx = usize::from(self.block.msg.req.arg[3]).into();
         }
 
         if TRACE {
